Project/Tollplaza_api/requirements.txt
autopy (0.51)
cycler (0.10.0)
Django (1.8)
django-braces (1.11.0)
django-filter (1.0.2)
django-oauth-toolkit (0.12.0)
django-tastypie (0.13.3)
djangorestframework (3.6.2)
EasyProcess (0.2.3)
functools32 (3.2.3.post2)
Markdown (2.6.8)
matplotlib (2.0.0)
mysql-python (1.2.5)
numpy (1.12.0)
oauthlib (2.0.1)
olefile (0.44)
opencv-python (3.2.0.6)
passlib (1.7.1)
Pillow (4.0.0)
pip (9.0.1)
pyBarcode (0.7)
pygame (1.9.3)
PyMySQL (0.7.9)
pyparsing (2.2.0)
pyping (0.0.5)
pypiwin32 (219)
pyscreenshot (0.4.2)
python-dateutil (2.6.0)
python-mimeparse (1.6.0)
pytz (2016.10)
requests (2.13.0)
selenium (3.3.0)
setuptools (32.0.0)
SimpleCV (1.3)
six (1.10.0)
virtualenv (15.1.0)
wheel (0.29.0)
